Removed the ``date_created`` field from models.
This field held the creation date of each model instance. If table records were created outside our admin interface, this field would be NULL. Consequently, if the user tried to edit the Model through the ModelAdmin, the ``date_created`` wouldn't be filled (since it is an edit of an *existing* instance) and thus there would be an error about the ``date_created`` being NULL. This situation would require manual fix through the database shell.